+++ src-tauri/src/infrastructure/database/mod.rs
@@ -1,4 +1,5 @@
 pub mod book_repository;
+pub mod bookmark_repository;
 pub mod bookshelf_repository;
 pub mod models;
 pub mod series_repository;